More refactoring on cli
