+++ rudof_cli/src/color_support.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone, PartialEq)]
+pub enum ColorSupport {
+    NoColor,
+    WithColor,
+}
