+++ rudof_cli/src/lib.rs
@@ -1,0 +1,57 @@
+// Current modules
+pub mod cli;
+pub mod color_support;
+pub mod data;
+pub mod data_format;
+pub mod dctap_format;
+pub mod dctap_result_format;
+pub mod input_convert_format;
+pub mod input_convert_mode;
+pub mod input_spec;
+pub mod mime_type;
+pub mod node;
+pub mod node_selector;
+pub mod output_convert_format;
+pub mod output_convert_mode;
+pub mod query;
+pub mod rdf_reader_mode;
+pub mod result_data_format;
+pub mod result_query_format;
+pub mod result_service_format;
+pub mod result_shacl_validation_format;
+pub mod result_shex_validation_format;
+pub mod result_validation_format;
+pub mod shacl_format;
+pub mod shapemap_format;
+pub mod shex;
+pub mod shex_format;
+pub mod show_mode;
+pub mod validation_mode;
+pub mod writer;
+
+pub use color_support::*;
+pub use dctap_result_format::*;
+pub use input_convert_format::*;
+pub use input_convert_mode::*;
+pub use input_spec::*;
+use iri_s::IriS;
+pub use output_convert_format::*;
+pub use output_convert_mode::*;
+pub use rdf_reader_mode::*;
+pub use result_data_format::*;
+pub use result_data_format::*;
+pub use result_query_format::*;
+pub use result_service_format::*;
+pub use result_shacl_validation_format::*;
+pub use result_shex_validation_format::*;
+pub use result_validation_format::*;
+pub use shacl_format::*;
+pub use shapemap_format::*;
+pub use shex::*;
+pub use shex_format::*;
+pub use show_mode::*;
+pub use validation_mode::*;
+
+fn base_convert(base: &Option<IriS>) -> Option<&str> {
+    base.as_ref().map(|iri| iri.as_str())
+}
