+++ rudof_cli/src/result_query_format.rs
@@ -1,0 +1,16 @@
+use clap::ValueEnum;
+use std::fmt::{Display, Formatter};
+
+#[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Debug)]
+#[clap(rename_all = "lower")]
+pub enum ResultQueryFormat {
+    Internal,
+}
+
+impl Display for ResultQueryFormat {
+    fn fmt(&self, dest: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
+        match self {
+            ResultQueryFormat::Internal => write!(dest, "internal"),
+        }
+    }
+}
