+++ rudof_cli/src/result_service_format.rs
@@ -1,0 +1,16 @@
+use clap::ValueEnum;
+use std::fmt::{Display, Formatter};
+
+#[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Debug)]
+#[clap(rename_all = "lower")]
+pub enum ResultServiceFormat {
+    Internal,
+}
+
+impl Display for ResultServiceFormat {
+    fn fmt(&self, dest: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
+        match self {
+            ResultServiceFormat::Internal => write!(dest, "internal"),
+        }
+    }
+}
