+++ rudof_cli/src/result_shex_validation_format.rs
@@ -1,0 +1,30 @@
+use clap::ValueEnum;
+use std::fmt::{Display, Formatter};
+
+#[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Debug)]
+#[clap(rename_all = "lower")]
+pub enum ResultShExValidationFormat {
+    Turtle,
+    NTriples,
+    RDFXML,
+    TriG,
+    N3,
+    NQuads,
+    Compact,
+    Json,
+}
+
+impl Display for ResultShExValidationFormat {
+    fn fmt(&self, dest: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
+        match self {
+            ResultShExValidationFormat::Turtle => write!(dest, "turtle"),
+            ResultShExValidationFormat::NTriples => write!(dest, "ntriples"),
+            ResultShExValidationFormat::RDFXML => write!(dest, "rdfxml"),
+            ResultShExValidationFormat::TriG => write!(dest, "trig"),
+            ResultShExValidationFormat::N3 => write!(dest, "n3"),
+            ResultShExValidationFormat::NQuads => write!(dest, "nquads"),
+            ResultShExValidationFormat::Compact => write!(dest, "compact"),
+            ResultShExValidationFormat::Json => write!(dest, "json"),
+        }
+    }
+}
