+++ rudof_cli/src/result_validation_format.rs
@@ -1,0 +1,30 @@
+use clap::ValueEnum;
+use std::fmt::{Display, Formatter};
+
+#[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Debug)]
+#[clap(rename_all = "lower")]
+pub enum ResultValidationFormat {
+    Turtle,
+    NTriples,
+    RDFXML,
+    TriG,
+    N3,
+    NQuads,
+    Compact,
+    Json,
+}
+
+impl Display for ResultValidationFormat {
+    fn fmt(&self, dest: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
+        match self {
+            ResultValidationFormat::Turtle => write!(dest, "turtle"),
+            ResultValidationFormat::NTriples => write!(dest, "ntriples"),
+            ResultValidationFormat::RDFXML => write!(dest, "rdfxml"),
+            ResultValidationFormat::TriG => write!(dest, "trig"),
+            ResultValidationFormat::N3 => write!(dest, "n3"),
+            ResultValidationFormat::NQuads => write!(dest, "nquads"),
+            ResultValidationFormat::Compact => write!(dest, "compact"),
+            ResultValidationFormat::Json => write!(dest, "json"),
+        }
+    }
+}
